Edit gitignore

+++ head/scheduler.h
@@ -9,6 +9,7 @@
 #include <list>
 #include <random>
 #include <cstdlib>
+#include <cmath>
 
 #include "scheduler.h"
 #include "paciente.h"
@@ -22,7 +23,7 @@ public:
     int iter;
     int param1;
     float paramProb;
-    int temp;
+    float temp;
     float multi;
     int iterTemp;
     int dias;
@@ -40,7 +41,7 @@ public:
     std::vector<int> capacidadMaquinas;
     float mejorSolucion;
 
-    Scheduler(int iteraciones, int parametro1, float probabilidad, int temperatura, float multiplicador, int iterTemperatura);
+    Scheduler(int iteraciones, int parametro1, float probabilidad, float temperatura, float multiplicador, int iterTemperatura);
     int leerInstancia(std::string instancia);
     bool compPrimeraCapacidad(int j, Paciente &paciente, std::vector<int> &capacidades);
     bool compCapacidad(int j, Paciente &paciente, std::vector<int> &capacidades);
